+++ src/main/java/ua/duikt/learning/java/pro/spring/entity/IssueHistory.java
@@ -16,9 +16,9 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class IssueHistory {
-    private Integer id;
-    private Integer issueId;
-    private Integer userId;
+    private Long id;
+    private Long issueId;
+    private Long userId;
     private String fieldChanged;
     private String oldValue;
     private String newValue;
